Fixe minor issues in reservoir sample tests

+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/differentialentropy/UnweightedReservoirSampleStateStrategy.java
@@ -95,7 +95,7 @@ public class UnweightedReservoirSampleStateStrategy
     @Override
     public double getTotalPopulationWeight()
     {
-        return (double) reservoir.getTotalPopulationCount();
+        return reservoir.getTotalPopulationCount();
     }
 
     @Override
